update status to use decimal

+++ programs/dsla/src/state/status_registry.rs
@@ -1,12 +1,5 @@
+use crate::state::utils::Decimal;
 use anchor_lang::prelude::*;
-
-/// struct used to generate the statuses for an SLA
-///
-///
-/// # Fields
-///
-///  * `statuses` - the status of a given period
-///
 #[account]
 pub struct StatusRegistry {
     pub statuses: Vec<Status>,
@@ -16,6 +9,6 @@ pub struct StatusRegistry {
 #[derive(AnchorSerialize, AnchorDeserialize, Debug, PartialEq, Clone)]
 pub enum Status {
     NotVerified,
-    Respected { value: u128 },
-    NotRespected { value: u128 },
+    Respected { value: Decimal },
+    NotRespected { value: Decimal },
 }
